+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Added: Sorting and aggregating of row ids and table names
 * Added: Option to execute a script from the shell
 * Fixed: Streaming property value input stream from data table
+* Fixed: Elasticsearch5 retry logic to sleep per request failure instead of batch of failures
 * Removed: ElasticSearch 1.x support
 
 # v3.0.4